+++ examples/adc_oneshot.rs
@@ -6,7 +6,7 @@ use std::time::Duration;
 
 #[cfg(not(any(feature = "adc-oneshot-legacy", esp_idf_version_major = "4")))]
 fn main() -> anyhow::Result<()> {
-    use esp_idf_hal::adc::attenuation::DB_11;
+    use esp_idf_hal::adc::attenuation::DB_12;
     use esp_idf_hal::adc::oneshot::config::AdcChannelConfig;
     use esp_idf_hal::adc::oneshot::*;
     use esp_idf_hal::peripherals::Peripherals;
@@ -22,7 +22,7 @@ fn main() -> anyhow::Result<()> {
     // configuring pin to analog read, you can regulate the adc input voltage range depending on your need
     // for this example we use the attenuation of 11db which sets the input voltage range to around 0-3.6V
     let config = AdcChannelConfig {
-        attenuation: DB_11,
+        attenuation: DB_12,
         ..Default::default()
     };